Cleanup
hope i dont broke anything.
